% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_dates}
\alias{parse_dates}
\title{Parse string dates into Dates}
\usage{
parse_dates(x)
}
\arguments{
\item{x}{A list of people together with availability dates (to have a beer!).
Dates could be expressed as date, name of week day or interval of dates.}
}
\value{
A list of date (in date format) for each person
}
\description{
Parse string dates into Dates
}
\examples{
beer_dates_string <- list( andrea = c("thursday", "friday"),
 federico = "2023-05-04",
 chiara = c("2023-05-03 / 2023-05-05", "2023-05-10"))

parse_dates(beer_dates_string)

}
