% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decide.R
\name{decide_happy_hour}
\alias{decide_happy_hour}
\title{Title Decide the best day gor a happy hour.}
\usage{
decide_happy_hour(x)
}
\arguments{
\item{x}{A list of date vectors}
}
\value{
The set of common dates
}
\description{
Title Decide the best day gor a happy hour.
}
\examples{
1+1
}
