% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decide.R
\name{decide_happy_hour}
\alias{decide_happy_hour}
\title{Decide the best day for happy hour}
\usage{
decide_happy_hour(x)
}
\arguments{
\item{x}{a list of Date vectors}
}
\value{
the set of common Date(s)
}
\description{
Decide the best day for happy hour
}
\examples{
1+1
}
