% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decide.R
\name{decide_happy_hour}
\alias{decide_happy_hour}
\title{Decide the best day for an happy hour}
\usage{
decide_happy_hour(x)
}
\arguments{
\item{x}{A list of Date vectors}
}
\value{
The set of common Date(s)
}
\description{
Decide the best day for an happy hour
}
\examples{
beer_dates_string <- list( andrea = c("thursday", "friday"),
 federico = "2023-05-04",
 chiara = c("2023-05-03 / 2023-05-05", "2023-05-10"))

dates <- parse_dates(beer_dates_string)
decide_happy_hour(dates)
}
