% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_dates}
\alias{parse_dates}
\title{Parse string dates into dates}
\usage{
parse_dates(x)
}
\arguments{
\item{x}{a list of character dates}
}
\value{
a list of Date vectors
}
\description{
Parse string dates into dates
}
\examples{
1 + 1
}
