% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_dates}
\alias{parse_dates}
\title{Parse string dates into Dates}
\usage{
parse_dates(x)
}
\arguments{
\item{x}{A list of character dates}
}
\value{
A list of Date vectors
}
\description{
Parse string dates into Dates
}
\examples{
1+1
}
